backend/requirements.txt
Flask==2.0.3
kafka-python==2.0.2
marshmallow==3.14.1
Werkzeug==2.0.3
flask-cors==3.0.10
flask-socketio==5.3.2
eventlet==0.33.3
gevent==21.12.0
gevent-websocket==0.10.1 
greenlet==1.1.2  